apps/backend/requirements.txt
fastapi
uvicorn
pydantic
langchain-openai
langchain-community
pinecone
transformers
sentence-transformers
huggingface-hub
fastembed
# Firecrawl is optional; install if using web crawling
dotenv
# For CORS support
httpx
# apps/backend/requirements.txt
pyjwt==2.8.0
python-multipart==0.0.9
python-jose==3.3.0
redis==5.0.3
fastapi-cache2==0.2.1
# For Prometheus metrics
prometheus_client
asyncpg
SQLAlchemy
alembic
# For PEP-8 formatting
black
flake8
# For PDF parsing
pdfplumber
# For file uploads
python-multipart
# For JWT authentication
PyJWT
# For scheduled anonymization
schedule
# For impact sorting
isort
#For testing
pytest<8.0.0
pytest-asyncio==0.21.0
# For rate limiting
psycopg2-binary
fastapi-limiter>=0.1.5
pytest-cov